requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-generativeai>=0.3.0
pandas>=1.3.5
matplotlib>=3.5.1
seaborn>=0.12.0
python-dotenv>=1.0.0
tqdm>=4.65.0
